+++ fabric20/src/main/java/com/vecoo/extralib/ExtraLib.java
@@ -1,6 +1,10 @@
 package com.vecoo.extralib;
 
+import com.vecoo.extralib.task.TaskTimer;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents;
+import net.minecraft.server.MinecraftServer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -10,9 +14,15 @@ public class ExtraLib implements ModInitializer {
 
     private static ExtraLib instance;
 
+    private MinecraftServer server;
+
     @Override
     public void onInitialize() {
         instance = this;
+
+        ServerLifecycleEvents.SERVER_STARTING.register(server -> this.server = server);
+        ServerTickEvents.END_SERVER_TICK.register(server -> TaskTimer.onServerTick());
+        ServerLifecycleEvents.SERVER_STOPPING.register(server -> TaskTimer.cancelAll());
     }
 
     public static ExtraLib getInstance() {
@@ -21,5 +31,9 @@ public class ExtraLib implements ModInitializer {
 
     public static Logger getLogger() {
         return LOGGER;
+    }
+
+    public MinecraftServer getServer() {
+        return instance.server;
     }
 }